requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
jinja2==3.1.5
itsdangerous==2.2.0

# Settings
pydantic==2.10.4
pydantic-settings==2.7.0

# Auth
pyotp==2.9.0
qrcode[pil]==8.0
PyJWT==2.10.1
cryptography==44.0.0

# Database
supabase==2.11.0
redis==5.2.1

# OKX Trading
python-okx==0.3.9

# AI
openai==1.59.9

# Technical Indicators
pandas==2.3.3
numpy==2.2.6
numba==0.61.2
pandas-ta==0.4.71b0

# HTTP / scraping
httpx==0.28.1
beautifulsoup4==4.12.3

# Logging
loguru==0.7.3

# Utils
python-dotenv==1.0.1
